🧹 Jules05: Technical debt cleanup — architectural harmonization
- Standardized Signal Mapping across models and adapters to follow ModelAction (0=HOLD, 1=BUY, 2=SELL).
- Consolidated LSTM architectures by moving LSTMAttentionModel to src/models/lstm_model.py.
- Remediated PR residue by replacing print() with logger in research scripts.
- Improved type safety by adding missing hints to PPOAgent and benchmark adapters.
- Updated docs/technical-debt/DEBT_LOG.md with latest audit findings.
