+++ src/research/hyperopt_walkforward.py
@@ -285,4 +285,4 @@ if __name__ == "__main__":
     )
 
     result = optimizer.run_optimization()
-    print(result.model_dump_json(indent=2))
+    logger.info("Optimization complete: %s", result.model_dump_json(indent=2))
